requirements.txt
streamlit
PyPDF2
python-dotenv
openai==0.28
os